_data/requirements.txt
certifi==2024.12.14
charset-normalizer==3.4.1
googledrivedownloader==0.4
idna==3.10
pillow==11.1.0
PyYAML==6.0.2
requests==2.32.3
urllib3==2.3.0
